added more data-argument tests
